requirements.txt
flask>=3.0
pandas>=2.0
openpyxl>=3.1
werkzeug>=3.0
python-dotenv>=1.0
